Update MsgResTonpy.trans
'that show DURING noon' - I'm not sure when the show is on. It is on right at 12pm or in the afternoon in general?  If so, we could change this to "at noon" or "in the afternoon".

What would you think of calling it the "Chibi-Sifter"?  'Sieve' and 'sifter' are synonyms, but I hear people say "sift" more often.