Update MsgResAca.trans
LOL, 'i' capitalized! Also I added in 'mishap' in that sentence meaning the character was diminishing/downplaying/trivalizing the impact of that crash.
Added grammer and some spelling errors. 
In line #85: Did you mean to say "My wishes for a bird chirp〜" like the character is pleading for that, or is it "My..., I wished for a bird chirp~"?
Line 194: According to some users 'accidently' is an old, correct variant. But to some other sources 'accidentally' is the only correct adverb.
Line 210: Is the bird meant to say she's not a cuckoo bird or is it an expression 'I have not gone cuckoo!' meaning she's not crazy?